+++ src/components/CarbonAd/inlineCarbonAd.astro
@@ -1,0 +1,19 @@
+---
+import CarbonAdReminder from './InlineCarbonAdReminder';
+import './InlineCarbonAd.css'
+
+const props = Astro.props
+---
+
+<main>
+    <div>
+        <script
+            key="carbonads"
+            async
+            id="_carbonads_js"
+            type="text/javascript"
+            src="//cdn.carbonads.com/carbon.js?serve=CKYIEK7E&placement=fabricjscom"
+        />
+        <CarbonAdReminder client:load />
+    <div>
+</main>
