+++ src/components/PageTitleDocs/PageTitle.astro
@@ -1,9 +1,9 @@
 ---
 import RealPageTitle from '@astrojs/starlight/components/PageTitle.astro';
-import CarbonAd from '../CarbonAd/InlineCarbonAd';
+import CarbonAd from '../CarbonAd/InlineCarbonAd.astro';
 
 const props = Astro.props
 ---
 
-<CarbonAd client:load />
+<CarbonAd />
 <RealPageTitle {...props} />
